Add spsc, fix tests

+++ code/src/test/java/com/github/srvaroa/queue/LinkedBackedQueueConcurrentTest.java
@@ -2,7 +2,7 @@ package com.github.srvaroa.queue;
 
 public class LinkedBackedQueueConcurrentTest extends QueueConcurrentTest {
     @Override
-    Queue<Integer> getInstance(int capacity) {
+    Queue<Long> getInstance(int capacity) {
         return new LinkedBackedQueue<>(capacity);
     }
 }
