+++ code/src/test/java/com/github/srvaroa/queue/SynchronizedQueueConcurrentTest.java
@@ -2,7 +2,7 @@ package com.github.srvaroa.queue;
 
 public class SynchronizedQueueConcurrentTest extends QueueConcurrentTest {
     @Override
-    Queue<Integer> getInstance(int capacity) {
+    Queue<Long> getInstance(int capacity) {
         return new SynchronizedQueue<>(capacity);
     }
 }
